requirements.txt
Flask==2.1.2
tensorflow==2.11.0
keras==2.11.0
pillow==9.3.0
numpy==1.23.3